Ensure functions passed as props exist
I really like how you defined the API-related functions in
`MessagesComponent`, then pass them down to the other components, then
have separate handlers that call the functions passed as props. That's
good practice. If these components were used other places, and someone
failed to pass the function as a prop, if statements like this would
prevent your JS from erroring out. You could even define a default
action in an else statement, or console something.

+++ pwnedhub/static/react/messages.js
@@ -61,8 +61,10 @@ var MessageForm = React.createClass({
     },
     handleFormSubmit(e) {
         e.preventDefault();
-        this.props.onMessageSubmit({message: this.state.message});
-        this.setState({ message: "" });
+        if (this.props.onMessageSubmit) {
+            this.props.onMessageSubmit({message: this.state.message});
+            this.setState({ message: "" });
+        }
     },
     onChange(e) {
         this.setState({ message: e.target.value });
@@ -105,7 +107,9 @@ var MessageList = React.createClass({
 var MessageDelete = React.createClass({
     handleDeleteClick(e) {
         e.preventDefault();
-        this.props.onDeleteMessage(this.props.message.id);
+        if (this.props.onDeleteMessage) {
+            this.props.onDeleteMessage(this.props.message.id);
+        }
     },
     render() {
         // prevent the componenet from mounting if it is not owned by the current user
